+++ lib/network_international_payment_sdk_platform_interface.dart
@@ -36,6 +36,7 @@ abstract class NetworkInternationalPaymentSdkPlatform extends PlatformInterface 
   Future<Map<dynamic, dynamic>?> startSavedCardPayment({
     required Map<String, dynamic> orderDetails,
     String? cvv,
+    Map<String, dynamic>? theme,
   }) {
     throw UnimplementedError('startSavedCardPayment() has not been implemented.');
   }
@@ -43,6 +44,7 @@ abstract class NetworkInternationalPaymentSdkPlatform extends PlatformInterface 
   Future<Map<dynamic, dynamic>?> startApplePay({
     required Map<String, dynamic> orderDetails,
     required Map<String, dynamic> applePayConfig,
+    Map<String, dynamic>? theme,
   }) {
     throw UnimplementedError('startApplePay() has not been implemented.');
   }